chore: remove unused files and clean up obsolete code
