--- app/db.js
@@ -1,13 +1,0 @@
-const mongoose = require('mongoose');
-
-const connectDB = async (uri) => {
-  try {
-    await mongoose.connect(uri);
-    console.log('MongoDB connected');
-  } catch (err) {
-    console.error('MongoDB connection failed:', err);
-    throw err;
-  }
-};
-
-module.exports = connectDB;
